# openapi.yaml
openapi: "3.0.3"
info:
  title: Blog management API
  version: "1.0"
  contact:
    name: Blog API
    url: api.blog.com
    email: support@blog.com
  license:
    name: Apache 2.0
    url: "something.com"
servers:
  - url: https://api.server.test/v1/{username}:{port}
    description: local Server
    variables:
      username:
        default: demo
        description: A user specific sub domain
      port:
        enum:
          - "4000"
          - "8080"
        default: "4000"
  - url: http://localhost:4000/api/v1
    description: Dev Server
...
paths:
...
  /articles:
    get:
      security:
        - bearerAuth: []
      operationId: getArticles
      description: Retrieve all published Article
      tags:
        - Article
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/limitParam"
        - $ref: "#/components/parameters/sortByParam"
        - $ref: "#/components/parameters/sortTypeParam"
        - $ref: "#/components/parameters/authorParam"
      responses:
        "200":
          description: successfully retrieved articles
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    $ref: "#/components/schemas/ArticleDTO"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/404"
      summary: Return published ariticle based on params
    post:
      security:
        - bearerAuth: []
      operationId: CreateArticle
      description: Create new article
      tags: [Article]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "JavaScript Tutorial"
                content:
                  type: string
                  example: First article content
                cover:
                  type: string
                  format: url
                  example: "https://s3.aws.com/cover/2.jpg"
                author:
                  type: integer
                  format: int32
                  example: 3
                status:
                  type: string
                  enum: [published, draft, removed, private]
              required: [title, content, cover, author]
      responses:
        "201":
          description: Article Created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  message:
                    type: string
                    example: Success
                  data:
                    $ref: "#/components/schemas/Article"
        "401":
          $ref: "#/components/responses/401"
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
      summary: take input and create new ariticle
...
components:
  schemas:
...
    Article:
      allOf:
        - type: object
          properties:
            id:
              type: string
              example: "68315e7f3cb9dea49ee067f8"
            title:
              type: string
              example: "JavaScript Tutorial"
            content:
              type: string
              example: First article content
            cover:
              type: string
              format: url
              example: "https://s3.aws.com/cover/2.jpg"
            author:
              $ref: "#/components/schemas/AuthorDTO"
            comments:
              type: array
              items:
                type: integer
                format: int32
            status:
              type: string
              enum: [published, draft, removed, private]
        - $ref: "#/components/schemas/TimeStamp"
...
    TimeStamp:
      type: object
      properties:
        createdAt:
          type: string
          example: "2025-04-09T18:59:40.964Z"
        updatedAt:
          type: string
          example: "2025-04-09T19:00:32.636Z"
    AuthorDTO:
      type: object
      properties:
        id:
          type: string
          example: "68315e7f3cb9dea49ee067f8"
        name:
          type: string
          example: Nazim Uddin
        avatar:
          type: string
          format: url
          example: "https://s3.aws.com/avatar/1.jpg"
    ArticleDTO:
      allOf:
        - type: object
          properties:
            id:
              type: string
              example: "68315e7f3cb9dea49ee067f8"
            title:
              type: string
              example: "JavaScript Tutorial"
            cover:
              type: string
              format: url
              example: "https://s3.aws.com/cover/2.jpg"
            author:
              $ref: "#/components/schemas/AuthorDTO"
            comments:
              type: integer
              example: 30
        - $ref: "#/components/schemas/TimeStamp"
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          example: 10
        totalResult:
          type: integer
          format: int32
          example: 45
        totalPage:
          type: integer
          format: int32
          example: 5
        currentPage:
          type: integer
          format: int32
          example: 2
        nextPage:
          type: integer
          format: int32
          example: 3
        prevPage:
          type: integer
          format: int32
          example: 1
  securitySchemes:
    api_key:
      type: apiKey
      name: api-key
      in: header
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basic:
      type: http
      scheme: basic
  parameters:
    pageParam:
      name: page
      description: Current page number
      in: query
      schema:
        type: integer
        example: 2
    limitParam:
      name: limit
      description: Maximum items that will return
      in: query
      schema:
        type: integer
        example: 10
    sortByParam:
      name: sort_by
      description: the property that will be use to sort
      in: query
      schema:
        type: string
        example: createdAt
    sortTypeParam:
      name: sort
      description: The sorting type. (assending or desending)
      in: query
      schema:
        type: string
        enum: [dsc, asc]
        default: dsc
    authorParam:
      name: author
      description: Return resource based on author
      in: query
      schema:
        type: string
        format: 68315e7f3cb9dea49ee067f8
...
  responses:
    "500":
      description: Server Error
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 500
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: Server Error
                  description:
                    type: string
                    example: Internal Server Error
    "404":
      description: Not Found
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 404
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: "Not Found"
                  description:
                    type: string
                    example: "Resource not found"
    "401":
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 401
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: Unauthorized
                  description:
                    type: string
                    example: You do not have permission to perfom this action
    "400":
      description: Bad request
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 400
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: Bad Request
                  description:
                    type: string
                    example: You provided wrong input
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          example: email
                        message:
                          type: string
                          example: email is required
                        in:
                          type: string
                          enum: [body, param]